+++ backend/SpringBoot_Service/src/main/java/com/postgresql/SpringBoot_Service/repo/UserRepo.java
@@ -6,7 +6,7 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import java.util.Optional;
 
 @RepositoryRestResource
-public interface UserRepo extends JpaRepository<Faculty_users, Integer> {
+public interface UserRepo extends JpaRepository<Faculty_users, Long> {
     Optional<Faculty_users> findByEmailAndTipUsera(String email, String tipUsera);
     Optional<Faculty_users> findByEmail(String email);
